// apps/countries-api-with-color-theme-switcher/src/utils/types.ts
export type AlteredCountriesApiResponse = Omit<CountriesApiResponse, 'borders'> & { borders: Border[] }

export type CountriesApiFields =
  | 'borders'
  | 'capital'
  | 'cca3'
  | 'currencies'
  | 'flags'
  | 'languages'
  | 'name'
  | 'nativeName'
  | 'population'
  | 'region'
  | 'subregion'
  | 'tld'

export interface CountriesApiResponse {
  name: Name
  currencies: Currencies
  languages: Languages
  tld: string[]
  cca3: string
  capital: string[]
  region: string
  subregion: string
  flags: Flags
  population: number
  borders: string[]
}

export interface Country {
  flag: Flags
  tld?: string
  cca3: string
  name: string
  region: string
  capital?: string
  subRegion: string
  population: string
  languages: string[]
  nativeName?: string
  currencies: string[]
  borders: Border[]
}

export interface Flags {
  png: string
  svg: string
  alt: string
}

export interface Border {
  name: string
  cca3: string
}

export interface Name {
  common: string
  official: string
  nativeName: Partial<Record<keyof Languages, OfficialCommon>>
}

export interface OfficialCommon {
  official: string
  common: string
}

export interface NameSymbol {
  name: string
  symbol: string
}

export interface Currencies {
  AED?: NameSymbol
  AFN?: NameSymbol
  ALL?: NameSymbol
  AMD?: NameSymbol
  ANG?: NameSymbol
  AOA?: NameSymbol
  ARS?: NameSymbol
  AUD?: NameSymbol
  AWG?: NameSymbol
  AZN?: NameSymbol
  BAM?: NameSymbol
  BBD?: NameSymbol
  BDT?: NameSymbol
  BGN?: NameSymbol
  BHD?: NameSymbol
  BIF?: NameSymbol
  BMD?: NameSymbol
  BND?: NameSymbol
  BOB?: NameSymbol
  BRL?: NameSymbol
  BSD?: NameSymbol
  BTN?: NameSymbol
  BWP?: NameSymbol
  BYN?: NameSymbol
  BZD?: NameSymbol
  CAD?: NameSymbol
  CDF?: NameSymbol
  CHF?: NameSymbol
  CKD?: NameSymbol
  CLP?: NameSymbol
  CNY?: NameSymbol
  COP?: NameSymbol
  CRC?: NameSymbol
  CUC?: NameSymbol
  CUP?: NameSymbol
  CVE?: NameSymbol
  CZK?: NameSymbol
  DJF?: NameSymbol
  DKK?: NameSymbol
  DOP?: NameSymbol
  DZD?: NameSymbol
  EGP?: NameSymbol
  ERN?: NameSymbol
  ETB?: NameSymbol
  EUR?: NameSymbol
  FJD?: NameSymbol
  FKP?: NameSymbol
  FOK?: NameSymbol
  GBP?: NameSymbol
  GEL?: NameSymbol
  GGP?: NameSymbol
  GHS?: NameSymbol
  GIP?: NameSymbol
  GMD?: NameSymbol
  GNF?: NameSymbol
  GTQ?: NameSymbol
  GYD?: NameSymbol
  HKD?: NameSymbol
  HNL?: NameSymbol
  HTG?: NameSymbol
  HUF?: NameSymbol
  IDR?: NameSymbol
  ILS?: NameSymbol
  IMP?: NameSymbol
  INR?: NameSymbol
  IQD?: NameSymbol
  IRR?: NameSymbol
  ISK?: NameSymbol
  JEP?: NameSymbol
  JMD?: NameSymbol
  JOD?: NameSymbol
  JPY?: NameSymbol
  KES?: NameSymbol
  KGS?: NameSymbol
  KHR?: NameSymbol
  KID?: NameSymbol
  KMF?: NameSymbol
  KPW?: NameSymbol
  KRW?: NameSymbol
  KWD?: NameSymbol
  KYD?: NameSymbol
  KZT?: NameSymbol
  LAK?: NameSymbol
  LBP?: NameSymbol
  LKR?: NameSymbol
  LRD?: NameSymbol
  LSL?: NameSymbol
  LYD?: NameSymbol
  MAD?: NameSymbol
  MDL?: NameSymbol
  MGA?: NameSymbol
  MKD?: NameSymbol
  MMK?: NameSymbol
  MNT?: NameSymbol
  MOP?: NameSymbol
  MRU?: NameSymbol
  MUR?: NameSymbol
  MVR?: NameSymbol
  MWK?: NameSymbol
  MXN?: NameSymbol
  MYR?: NameSymbol
  MZN?: NameSymbol
  NAD?: NameSymbol
  NGN?: NameSymbol
  NIO?: NameSymbol
  NOK?: NameSymbol
  NPR?: NameSymbol
  NZD?: NameSymbol
  OMR?: NameSymbol
  PAB?: NameSymbol
  PEN?: NameSymbol
  PGK?: NameSymbol
  PHP?: NameSymbol
  PKR?: NameSymbol
  PLN?: NameSymbol
  PYG?: NameSymbol
  QAR?: NameSymbol
  RON?: NameSymbol
  RSD?: NameSymbol
  RUB?: NameSymbol
  RWF?: NameSymbol
  SAR?: NameSymbol
  SBD?: NameSymbol
  SCR?: NameSymbol
  SDG?: NameSymbol
  SEK?: NameSymbol
  SGD?: NameSymbol
  SHP?: NameSymbol
  SLE?: NameSymbol
  SOS?: NameSymbol
  SRD?: NameSymbol
  SSP?: NameSymbol
  STN?: NameSymbol
  SYP?: NameSymbol
  SZL?: NameSymbol
  THB?: NameSymbol
  TJS?: NameSymbol
  TMT?: NameSymbol
  TND?: NameSymbol
  TOP?: NameSymbol
  TRY?: NameSymbol
  TTD?: NameSymbol
  TVD?: NameSymbol
  TWD?: NameSymbol
  TZS?: NameSymbol
  UAH?: NameSymbol
  UGX?: NameSymbol
  USD?: NameSymbol
  UYU?: NameSymbol
  UZS?: NameSymbol
  VES?: NameSymbol
  VND?: NameSymbol
  VUV?: NameSymbol
  WST?: NameSymbol
  XAF?: NameSymbol
  XCD?: NameSymbol
  XOF?: NameSymbol
  XPF?: NameSymbol
  YER?: NameSymbol
  ZAR?: NameSymbol
  ZMW?: NameSymbol
  ZWL?: NameSymbol
}

export interface Languages {
  afr?: string
  amh?: string
  ara?: string
  arc?: string
  aym?: string
  aze?: string
  bar?: string
  bel?: string
  ben?: string
  ber?: string
  bis?: string
  bjz?: string
  bos?: string
  bul?: string
  bwg?: string
  cal?: string
  cat?: string
  ces?: string
  cha?: string
  ckb?: string
  cnr?: string
  crs?: string
  dan?: string
  deu?: string
  div?: string
  dzo?: string
  ell?: string
  eng?: string
  est?: string
  eus?: string
  fao?: string
  fas?: string
  fij?: string
  fil?: string
  fin?: string
  fra?: string
  gil?: string
  glc?: string
  gle?: string
  glv?: string
  grn?: string
  gsw?: string
  hat?: string
  heb?: string
  her?: string
  hgm?: string
  hif?: string
  hin?: string
  hmo?: string
  hrv?: string
  hun?: string
  hye?: string
  ind?: string
  isl?: string
  ita?: string
  jam?: string
  jpn?: string
  kal?: string
  kat?: string
  kaz?: string
  kck?: string
  khi?: string
  khm?: string
  kin?: string
  kir?: string
  kon?: string
  kor?: string
  kwn?: string
  lao?: string
  lat?: string
  lav?: string
  lin?: string
  lit?: string
  loz?: string
  ltz?: string
  lua?: string
  mah?: string
  mey?: string
  mfe?: string
  mkd?: string
  mlg?: string
  mlt?: string
  mon?: string
  mri?: string
  msa?: string
  mya?: string
  nau?: string
  nbl?: string
  ndc?: string
  nde?: string
  ndo?: string
  nep?: string
  nfr?: string
  niu?: string
  nld?: string
  nno?: string
  nob?: string
  nor?: string
  nrf?: string
  nso?: string
  nya?: string
  nzs?: string
  pap?: string
  pau?: string
  pih?: string
  pol?: string
  por?: string
  pov?: string
  prs?: string
  pus?: string
  que?: string
  rar?: string
  roh?: string
  ron?: string
  run?: string
  rus?: string
  sag?: string
  sin?: string
  slk?: string
  slv?: string
  smi?: string
  smo?: string
  sna?: string
  som?: string
  sot?: string
  spa?: string
  sqi?: string
  srp?: string
  ssw?: string
  swa?: string
  swe?: string
  tam?: string
  tet?: string
  tgk?: string
  tha?: string
  tir?: string
  tkl?: string
  toi?: string
  ton?: string
  tpi?: string
  tsn?: string
  tso?: string
  tuk?: string
  tur?: string
  tvl?: string
  ukr?: string
  urd?: string
  uzb?: string
  ven?: string
  vie?: string
  xho?: string
  zdj?: string
  zho?: string
  zib?: string
  zul?: string
}